requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
ultralytics>=8.1.0
opencv-python-headless>=4.9.0
httpx>=0.27.0
openai>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
